requirements.txt
gitFlask==0.10.1
Flask-Bootstrap==3.2.0.2
Flask-Moment==0.4.0
Flask-Script==2.0.5
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
WTForms==2.0.1
Werkzeug==0.9.6
itsdangerous==0.24
pymongo==2.7.2
python-firebase==1.2
requests==2.4.3
wsgiref==0.1.2
